+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## x.x.x yyyy-yy-yy
 ### PaymentSheet
 * [Added] Ability to update saved cards when using CustomerSessions (private preview)
+* [Added] Added payment method options setup future usage to deferred API.
 
 ## 24.11.1 2025-04-14
 ### PaymentSheet